refactoring commands registration
